#3 - Added Stalwart (Minor Fixes Needed)
Added Stalwart and tweaked it so that it mostly works properly; still some minor issues but the skill is in a functional state - will fix at a later date.

Minor issues - Weapons/skills that inflict status effects will wear off at the start of the next turn instead of after combat.