#9 - Quick Riposte, Fail + Gale, Solar Brace, Lunar Brace
Added Quick Riposte, Fail + Gale, Solar Brace, Lunar Brace

Changed Quick Riposte to only work if the user is at full HP, can change this back to >=50% if needed

Attempted to implement WeaponInsight - seems to be a problem where the ROM just ignores the skill entirely after building? I've definitely put all the IDs and pointers in the right places, and the problem doesn't seem to be with the code. Will leave it for now.